Test&Simulations commit
Remove Windows and iwrites to do simulations with BBones

+++ src/RadioSystem/ProcessingCoef.h
@@ -15,16 +15,16 @@ class ProcessingCoef {
 public:
 	ProcessingCoef();
 	ProcessingCoef(float Pt_exp_coef);
-	float getProcessingTimeCoef();
-	float setAlphad(float Pdpx, float Pdip, float Pe, float Pm);
-	float getAlphad();
-	void AddObservation(float processingtime, int Npixels, int Nip, double alphad);
+	double getProcessingTimeCoef();
+	double setAlphad(double Pdpx, double Pdip, double Pe, double Pm);
+	double getAlphad();
+	void AddObservation(double processingtime, int Npixels, int Nip, double alphad);
 	void setOverlap(double overlap_normalized, bool double_overlap);
 private:
-	float Ptcoef;
+	double Ptcoef;
 	int pt_samples;
 	float Pt_exp_coef_;
-	float alpha_d_;
+	double alpha_d_;
 	double overlapNpixels;
 };
